+++ tests/testthat/test-strat-sample.R
@@ -1,48 +1,51 @@
 context("test-strat-sample")
 
-data(densitygrid)
+# make expected results
 col <- 'density'
-pts <- strat_sample(densitygrid, 10, col = col, returnDT = TRUE)
+pts <- sample_ct(densitygrid, 10, type = 'random',
+										col = col, returnDT = TRUE)
 
-
+# tests
 test_that("basics", {
-	expect_error(strat_sample(densitygrid, 10, col = 'potato'),
+	expect_error(sample_ct(densitygrid, 10, type = 'random', col = 'potato'),
 							 error = 'strata column not found in x')
 
 	densitygrid$potato <- NA
-	expect_error(strat_sample(densitygrid, 10, col = 'potato'),
+	expect_error(sample_ct(densitygrid, 10, type = 'random', col = 'potato'),
 							 error = 'no strata found')
+
+	expect_error(sample_ct(densitygrid, 10, col = col),
+							 error = 'type must be provided. either "regular" or "random"')
+
 })
 
-
-
 test_that("with data.table return", {
-	pts <- strat_sample(densitygrid, 10,
+	pts <- sample_ct(densitygrid, 10, type = 'random',
 											col = col, returnDT = TRUE)
 
 	expect_equivalent(unique(pts[[col]]),
 										unique(densitygrid[[col]]))
 
 
-	expect_true('data.table' %in% class(pts))
+	expect_true(inherits(pts, 'data.table'))
 
 })
 
 test_that("with sf return", {
-	pts <- strat_sample(densitygrid, 10,
+	pts <- sample_ct(densitygrid, 10, type = 'random',
 											col = col, returnDT = FALSE)
 
 	expect_equivalent(unique(pts[[col]]),
 										unique(densitygrid[[col]]))
 
-	expect_true('sf' %in% class(pts))
+	expect_true(inherits(pts, 'sf'))
 })
 
 
 test_that("type of col is well handled", {
 	# factor
 	densitygrid$tryfactor <- as.factor(densitygrid$density)
-	pts <- strat_sample(densitygrid, 10,
+	pts <- sample_ct(densitygrid, 10, type = 'random',
 											col = 'tryfactor', returnDT = FALSE)
 
 	expect_equivalent(unique(pts[['tryfactor']]),
@@ -50,7 +53,7 @@ test_that("type of col is well handled", {
 
 	# numeric
 	densitygrid$trynum <- sample(1:3, size = 4, replace = TRUE)
-	pts <- strat_sample(densitygrid, 10,
+	pts <- sample_ct(densitygrid, 10, type = 'random',
 											col = 'trynum', returnDT = FALSE)
 
 	expect_equivalent(unique(pts[['trynum']]),
@@ -58,7 +61,7 @@ test_that("type of col is well handled", {
 
 	# boolean
 	densitygrid$trybool <- sample(c(TRUE, FALSE), size = 4, replace = TRUE)
-	pts <- strat_sample(densitygrid, 10,
+	pts <- sample_ct(densitygrid, 10, type = 'random',
 											col = 'trybool', returnDT = FALSE)
 
 	expect_equivalent(unique(pts[['trybool']]),